[refactor] : 북마크 도메인 분리
